fix: 1. publish_service 使用类绑定属性替换字符串 selectinload
SQLAlchemy 不接受字符串形式的 loader option 属性名，
改用 Tool.mcp_server 和 Skill.tools 类绑定属性。

2. agent_publish agent_publishes sqlite 下BigInteger PRIMARY KEY 不等同于 ROWID，不会 auto-increment；其他表的迁移用的是sa.Integer()

+++ backend/src/models/agent_publish.py
@@ -18,7 +18,7 @@ class AgentPublish(Base):
         Index("ix_agent_publishes_agent_id_is_active", "agent_id", "is_active"),
     )
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     agent_id: Mapped[int] = mapped_column(BigInteger, ForeignKey("agents.id"), nullable=False)
     version: Mapped[int] = mapped_column(Integer, nullable=False, default=1)
     config_snapshot: Mapped[dict] = mapped_column(JSON, nullable=False)
